# ARM/NEON optimized crypto routines from kernel 3.13.y