… arch/arm/crypto/aes-armv4.S
… arch/arm/crypto/aesbs-core.S
… arch/arm/crypto/sha1-armv4-large.S
… crypto/ablk_helper.c
… include/crypto/ablk_helper.h