I ported arm/neon optimized crypto routines for AES and SHA1. The average speed-up is close to 50 per cent.

Cryptsetup test with general C routines
   Algorithm | Key | Encryption |  Decryption
     aes-cbc  | 128b |   18.9 MiB/s  |  20.2 MiB/s
     aes-cbc  | 256b |   13.9 MiB/s  |   15.2 MiB/s
     aes-xts  | 256b  |  18.8 MiB/s  |  19.8 MiB/s
     aes-xts  | 512b  |  14.5 MiB/s  |  15.0 MiB/s
 
Cryptsetup test w/ optimized routines
   Algorithm | Key |  Encryption |  Decryption
     aes-cbc |  128b  |  27.7 MiB/s  |  29.8 MiB/s
     aes-cbc |  256b  |  20.8 MiB/s  |  22.4 MiB/s
     aes-xts  | 256b  |  30.7 MiB/s  | 29.3 MiB/s
     aes-xts  | 512b  |  23.6 MiB/s  |  22.1 MiB/s

Default kernel config files need to be updated to enable these routines.